backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nfter"
version = "0.1.0"
description = "NFTer backend for generating Naruto-style anime avatars"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-multipart>=0.0.5",
    "replicate>=0.8.0",
    "python-dotenv>=0.19.0",
    "pydantic>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-asyncio>=0.15.0",
    "pytest-env>=0.8.1",
    "black>=21.7b0",
    "isort>=5.9.0",
    "mypy>=0.910",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true 